refactor: :recycle: reorganizando o codigo
