+++ src/utils/password-utils.ts
@@ -1,0 +1,12 @@
+import bcrypt from 'bcrypt';
+
+const SALT_ROUNDS = 10;
+
+export const hashPassword = async (password: string) => {
+  const salt = await bcrypt.genSalt(SALT_ROUNDS);
+  return await bcrypt.hash(password, salt);
+};
+
+export const comparePassword = async (password: string, hash: string) => {
+  return await bcrypt.compare(password, hash);
+};
